Extract CLI::FilterInput for reuse
Move explain's filter-input plumbing (collect_filters, read_raw_bpf,
raw_bpf_file?, dump_filters, executable?, source_name) into a FilterInput
mixin so the upcoming audit command reads the same inputs (BPF file, stdin,
executable, --pid) without duplication.
