+++ lib/seccomp-tools/cli/explain.rb
@@ -1,18 +1,16 @@
 # frozen_string_literal: true
 
 require 'seccomp-tools/cli/base'
-require 'seccomp-tools/cli/dumpable'
+require 'seccomp-tools/cli/filter_input'
 require 'seccomp-tools/disasm/disasm'
-require 'seccomp-tools/dumper'
 require 'seccomp-tools/explain'
 require 'seccomp-tools/logger'
-require 'seccomp-tools/util'
 
 module SeccompTools
   module CLI
     # Handle 'explain' command.
     class Explain < Base
-      include Dumpable
+      include FilterInput
 
       # Summary of this command.
       SUMMARY = 'Summarize a seccomp filter as a per-action policy.'
@@ -75,77 +73,6 @@ module SeccompTools
           insts = SeccompTools::Disasm.to_bpf(raw, arch).map(&:inst)
           output { SeccompTools::Explain.new(insts, arch:, source: label).summarize.to_s }
         end
-      end
-
-      private
-
-      # Resolves the input into an array of +[raw_bpf, arch, source]+ tuples, empty when there is
-      # nothing to explain (help shown, or an error was logged).
-      #
-      # The input is one of three kinds:
-      # * a running process, when +--pid+ is given;
-      # * a raw BPF file (or stdin), when the positional argument is not an executable;
-      # * a command to run and trace - either +-c+, or a positional executable.
-      # @return [Array<Array(String, Symbol, String?)>]
-      def collect_filters
-        # -c/--sh-exec and --pid take precedence over a positional BPF file or executable.
-        option[:ifile] = argv.shift unless option[:command] || option[:pid]
-        warn_ignored_arguments
-
-        return dump_filters(command: nil, pid: option[:pid], source: "pid #{option[:pid]}") if option[:pid]
-
-        command = option[:command] || option[:ifile]
-        if command.nil? # nothing to explain
-          CLI.show(parser.help)
-          return []
-        end
-        return read_raw_bpf if raw_bpf_file?
-
-        dump_filters(command:, pid: nil, source: command)
-      end
-
-      # Reads the positional file (or stdin) as a raw BPF blob, logging an error instead of
-      # crashing when it cannot be read.
-      # @return [Array<Array(String, Symbol, String?)>]
-      def read_raw_bpf
-        [[input, option[:arch], source_name(option[:ifile])]]
-      rescue SystemCallError => e
-        Logger.error(e.message)
-        []
-      end
-
-      # Should the input be read directly as a raw BPF blob, rather than run as a command? True when
-      # no +-c+ was given and the positional argument is not an executable (a plain file or stdin).
-      # @return [Boolean]
-      def raw_bpf_file?
-        !option[:command] && !executable?(option[:ifile])
-      end
-
-      # Dumps filters from a command or pid and labels each with +source+.
-      # @return [Array<Array(String, Symbol, String?)>]
-      #   The filter tuples, empty when dumping is unsupported or nothing was installed.
-      def dump_filters(command:, pid:, source:)
-        return [] unless dumping_supported?
-
-        dump_seccomp(command:, pid:, limit: option[:limit], timeout: option[:timeout]) do |bpf, arch|
-          [bpf, arch || option[:arch], source]
-        end
-      end
-
-      # Is +file+ an ELF executable to run, rather than a raw BPF blob or stdin to read?
-      # @param [String?] file
-      #   The path to check. +nil+ (no argument) and +-+ (stdin) are not executables.
-      # @return [Boolean]
-      def executable?(file)
-        file && file != '-' && Util.elf?(file)
-      end
-
-      # The label shown in the policy header for +file+; +<STDIN>+ when reading from stdin.
-      # @param [String] file
-      #   The input path, or +-+ for stdin.
-      # @return [String]
-      def source_name(file)
-        file == '-' ? '<STDIN>' : file
       end
     end
   end
